move ellipse

+++ classmain/src/MyCircle.java
@@ -3,7 +3,7 @@ import java.awt.geom.Ellipse2D;
 
 
 public class MyCircle extends MyEllipse{
-	public MyCircle(double x, double y, double R){
+	public MyCircle(int x, int y, double R){
 		super(x, y, R, R);
 	}
 }
